version 1.0.2.10 released
.ADD Planning Action History Function
.Import excel and auto insert/update to DB
.Print Item Data List
.carton auto out when coil top/tube coil top production in
.packing A: auto convert from piece to meter when production out
.all joined sub material calculate with join qty

+++ FactoryManagementSoftware/FactoryManagementSoftware/BLL/joinBLL.cs
@@ -10,7 +10,7 @@ namespace FactoryManagementSoftware.BLL
     {
         public string join_parent_code { get; set; }
         public string join_child_code { get; set; }
-        public int join_qty { get; set; }
+        public float join_qty { get; set; }
         
         public int join_max { get; set; }
         public int join_min { get; set; }
